Various additions and updates
